+++ src/main/java/racing/domain/car/CarMoveBehavior.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain.car;
 
 @FunctionalInterface
 public interface CarMoveBehavior {
