+++ src/main/java/racing/domain/car/RacingCars.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain.car;
 
 import java.util.*;
