+++ src/test/java/racing/RacingCarsTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import racing.domain.car.RacingCars;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
