+++ src/test/java/racing/RandomNumberTest.java
@@ -3,6 +3,7 @@ package racing;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racing.utils.RandomNumber;
 
 import static org.assertj.core.api.Assertions.assertThat;
